fix #29: Minor cleanup in kernel_tests
Finishing off the kernel tests there.
It does not look particularly nice, but it gets the job done.
Did a lot of copy pasting code to keep the testing quick.
The tests should be rewritten in the future, because this is not really
good code, but I want to start with the string puzzle.